Revert "Revert "(ugui) Move all files to deps folder""
This reverts commit 87a695d.
